+++ packages/orq-rc/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.10.0-rc.68",
+  sdkVersion: "4.10.0-rc.69",
   genVersion: "2.892.5",
-  userAgent: "speakeasy-sdk/typescript 4.10.0-rc.68 2.892.5 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.10.0-rc.69 2.892.5 2.0 @orq-ai/node",
 } as const;