+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-02T08:11:24.200Z",
+    "2026-06-02T09:50:17.688Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {